// src/shared/interfaces/WhyUsItem.ts
// Generated by https://quicktype.io

export interface WhyUsMenu {
  sneaks: Sneaks;
  team: Team;
  updates: Updates;
}

export interface Sneaks {
  title: string;
  items: SneaksItem[];
}

export interface SneaksItem {
  title: string;
  description: string;
  image: string;
}

export interface Team {
  title: string;
  items: TeamItem[];
}

export interface TeamItem {
  title: string;
  member: Member[];
}

export interface Member {
  look: string;
  name: string;
  profession_text: string;
}

export interface Updates {
  title: string;
  items: UpdatesItem[];
}

export interface UpdatesItem {
  author: string;
  author_look: string;
  date: string;
  update_text: string;
  image: string;
  tags: Tag[];
}

export interface Tag {
  name: string;
}